Move GPT config to separate file
